+++ app/(home)/page.tsx
@@ -1,0 +1,11 @@
+import ProfileBanner from "@/components/profile/profile-banner";
+
+const HomePage = () => {
+  return (
+    <>
+      <ProfileBanner />
+    </>
+  );
+};
+
+export default HomePage;
