--- app/page.tsx
@@ -1,11 +1,0 @@
-import Profile from "@/components/profile/profile-banner";
-
-const HomePage = () => {
-  return (
-    <>
-      <Profile />
-    </>
-  );
-};
-
-export default HomePage;
