--- components/posts-container.tsx
@@ -1,15 +1,0 @@
-import PostCard from "./posts/post-card";
-
-const PostsContainer = () => {
-  return (
-    <div className="mt-3">
-      <PostCard />
-      <PostCard />
-      <PostCard />
-      <PostCard />
-      <PostCard />
-    </div>
-  );
-};
-
-export default PostsContainer;
