+++ content/blog/hello-world.mdx
@@ -1,7 +1,7 @@
 ---
 title: Hello World
 description: This is our first post
-date: 2024-03-06
+date: 2024-03-07
 published: true
 ---
