+++ content/blog/second-world.mdx
@@ -1,0 +1,10 @@
+---
+title: Second World
+description: This is our second post
+date: 2024-03-12
+published: true
+---
+
+# Hello World
+
+Welcome to my blog
